# note-keeper-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/auth/login": {
            "post": {
                "description": "login",
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login",
                        "name": "loginForm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payload.LoginForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/presenter.Error400Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/presenter.Error404Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenter.Error500Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "payload.LoginForm": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "presenter.AuthResult": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "refreshToken": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "presenter.Error400Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Error Message"
                },
                "results": {},
                "status": {
                    "type": "string",
                    "example": "400"
                }
            }
        },
        "presenter.Error404Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Error Message"
                },
                "results": {},
                "status": {
                    "type": "string",
                    "example": "500"
                }
            }
        },
        "presenter.Error500Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Error Message"
                },
                "results": {},
                "status": {
                    "type": "string",
                    "example": "500"
                }
            }
        },
        "presenter.LoginResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "SUCCESS"
                },
                "results": {
                    "$ref": "#/definitions/presenter.AuthResult"
                },
                "status": {
                    "type": "string",
                    "example": "200"
                }
            }
        }
    }
}